fix(desktop): unify timeline signal icons and divider styling
